pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["chaewon-huh <cw.huh@postech.ac.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.0.0"
openai = "^1.58.1"
tiktoken = "^0.8.0"
streamlit = "^1.41.1"
python-dotenv = "^1.0.1"
qdrant-client = "^1.12.2"
flagembedding = "^1.3.3"
docx = "^0.2.4"
python-docx = "^1.1.2"
pdfplumber = "^0.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
